requirements.txt
aiogram==3.3.0
loguru==0.7.2
peewee==3.17.0
python-dotenv==1.0.1

gspread==6.0.1
oauth2client==4.1.3
google-api-python-client==2.116.0